% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acquire}
\alias{acquire}
\title{Launches Momacs}
\usage{
acquire(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{\code{logical} whether to launch in a browser (\code{TRUE}) by default)}
}
\value{
void
}
\description{
Launches Momacs
}
\examples{
\dontrun{
acquire()
}
}
