% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momacs-package.R
\docType{package}
\name{Momacs-package}
\alias{Momacs-package}
\title{Momacs: Morphometric Digitization Application}
\description{
Interactive Shiny application for digitizing morphometric data from images. Supports landmarks, polylines, polygons, curves, and outlines with export to JSON format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MomX/Momacs}
  \item \url{https://momx.github.io/Momacs/}
  \item Report bugs at \url{https://github.com/MomX/Momacs/issues}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com}

}
\keyword{internal}
